Organize files, options menu close outside
