+++ app/src/components/base/Container.tsx
@@ -2,10 +2,10 @@ import { useContext } from "react"
 import styles from "@/styles/base/Base.module.scss"
 import Head from "next/head"
 import lazy from "next/dynamic"
-import SidebarMenu from "./navigation/SidebarMenu"
+import SidebarMenu from "../navigation/Sidebar/SidebarMenu"
 import { SidebarContext } from "@/utils/Contexts"
 
-const MarginFiller = lazy(() => import("./MarginFiller.client"), {
+const MarginFiller = lazy(() => import("../MarginFiller.client"), {
   ssr: false
 })
