+++ app/src/components/base/Navbar/Navbar.tsx
@@ -1,8 +1,8 @@
 import { useRouter } from "next/router"
-import OptionsMenu from "../navigation/OptionsMenu"
-import SearchBox from "./SearchBox"
-import NavbarRoot from "./NavbarRoot"
-import LogoNav from "./LogoNav"
+import { NavbarRoot } from "./NavbarRoot"
+import { LogoNav } from "./LogoNav"
+import SearchBox from "../Searchbox/SearchBox"
+import OptionsMenu from "../../navigation/Options/OptionsMenu"
 
 export default function Navbar() {
   const router = useRouter()
