+++ app/src/components/base/Navbar/index.tsx
@@ -1,0 +1,3 @@
+export { default } from "./Navbar"
+export { NavbarRoot } from "./NavbarRoot"
+export { LogoNav } from "./LogoNav"
