+++ app/src/components/navigation/Options/index.tsx
@@ -1,0 +1,2 @@
+export { default } from "./OptionsMenu"
+export { OptionsItem } from "./OptionsItem"
