+++ app/src/components/navigation/Sidebar/index.tsx
@@ -1,0 +1,2 @@
+export { default } from "./SidebarMenu"
+export { SidebarItem } from "./SidebarItem"
