+++ app/src/pages/404.tsx
@@ -1,4 +1,4 @@
-import Container from "@/components/Container";
+import Container from "@/components/base/Container";
 
 export default function NotFound() {
   return(
