+++ app/src/pages/about.mdx
@@ -1,5 +1,5 @@
 import styles from "@/styles/pages/About.module.scss"
-import Container from "@/components/Container"
+import Container from "@/components/base/Container"
 
 # About _Searchpets!_
