+++ app/src/pages/characters.tsx
@@ -1,4 +1,4 @@
-import Container from "@/components/Container"
+import Container from "@/components/base/Container"
 
 export default function characters() {
   return <Container wrap>Characters page</Container>
